+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.2] - 2020-05-13
+### Added
+- This project is now a Homebridge verified plugin!
+### Fixed
+- Handle case where accessory.context.device is missing
+
 ## [2.0.1] - 2020-05-04
 ### Added
 - This change log!